Move HandlerAttributes and construct AbstractHandlerAttributes
The HandlerAttributes are not annotation specific, so should not reside
in the annotation package. Furthermore, we can benefit from an
AbstractHandlerAttributes, to deduplicate code in the Annotated- and
GenericHandlerAttributes class'

#1621
