+++ messaging/src/main/java/org/axonframework/messaging/annotation/MessageHandlingMember.java
@@ -163,8 +163,8 @@ public interface MessageHandlingMember<T> {
      * hold a value referencing the {@code attributeKey}, an {@link Optional#empty()} is returned. Otherwise a non-empty
      * {@link Optional} containing the attribute will be provided.
      * <p>
-     * When using the method, consider checking the {@link HandlerAttributes} for a list of common handler
-     * attributes.
+     * When using the method, consider checking the {@link org.axonframework.messaging.HandlerAttributes} for a list of
+     * common handler attributes.
      *
      * @param attributeKey the key to retrieve an attribute for
      * @return a non-empty {@link Optional} of the attribute referencing the given {@code attributeKey}. Otherwise an
