+++ messaging/src/test/java/org/axonframework/messaging/annotation/AnnotatedHandlerAttributesTest.java
@@ -6,6 +6,7 @@ import org.axonframework.deadline.DeadlineMessage;
 import org.axonframework.eventhandling.AllowReplay;
 import org.axonframework.eventhandling.EventHandler;
 import org.axonframework.eventhandling.EventMessage;
+import org.axonframework.messaging.HandlerAttributes;
 import org.axonframework.messaging.Message;
 import org.axonframework.messaging.interceptors.ExceptionHandler;
 import org.axonframework.messaging.interceptors.MessageHandlerInterceptor;
